feat: add new vite-dev-server package based on object API & bundling (#20925)
* improve error experience for cy-in-cy

* feat: Adding vite-dev-server-fresh w/ react tests against 2.8.6 & 2.9.1

* add system tests for vite-dev-server-fresh

* cleanup a bit

* Add snapshots

* Update system-tests/project-fixtures/react/cypress-vite.config.ts


* add comments

* Update npm/vite-dev-server-fresh/src/devServer.ts


* fix: update logLevel for snapshots

* fix logging of vite inspect URL

* add experimentalSessionSupport per @brian-mann suggestion to clear state between tests

+++ packages/frontend-shared/cypress/e2e/e2ePluginSetup.ts
@@ -114,7 +114,7 @@ async function makeE2ETasks () {
 
   const launchpadPort = await makeGraphQLServer()
 
-  const __internal_scaffoldProject = async (projectName: string) => {
+  const __internal_scaffoldProject = async (projectName: string, isRetry = false): Promise<string> => {
     if (fs.existsSync(Fixtures.projectPath(projectName))) {
       Fixtures.removeProject(projectName)
     }
@@ -128,12 +128,16 @@ async function makeE2ETasks () {
     try {
       await scaffoldProjectNodeModules(projectName)
     } catch (e) {
+      if (isRetry) {
+        throw e
+      }
+
       // If we have an error, it's likely that we don't have a lockfile, or it's out of date.
       // Let's run a quick "yarn" in the directory, kill the node_modules, and try again
       await execa('yarn', { cwd: Fixtures.projectFixturePath(projectName), stdio: 'inherit', shell: true })
       await fs.remove(path.join(Fixtures.projectFixturePath(projectName), 'node_modules'))
 
-      await scaffoldProjectNodeModules(projectName, true)
+      return await __internal_scaffoldProject(projectName, true)
     }
 
     scaffoldedProjects.add(projectName)
